macOS: move converters into `converters` directory
