+++ lib/macos/converters/window_effect_to_material_converter.dart
