+++ lib/macos/visual_effect_view_properties.dart
@@ -1,6 +1,6 @@
 import 'package:flutter_acrylic/flutter_acrylic.dart';
-import 'package:flutter_acrylic/macos/blur_view_state_to_visual_effect_view_state_converter.dart';
-import 'package:flutter_acrylic/macos/window_effect_to_material_converter.dart';
+import 'package:flutter_acrylic/macos/converters/blur_view_state_to_visual_effect_view_state_converter.dart';
+import 'package:flutter_acrylic/macos/converters/window_effect_to_material_converter.dart';
 
 import 'package:macos_window_utils/macos/visual_effect_view_properties.dart'
     as MacOSWindowUtilsVisualEffectSubviewProperties;
